Refactor into package and unify ProteinDB
