+++ protein_db/__init__.py
@@ -1,0 +1,20 @@
+"""Core classes and utilities for the Protein-DB package."""
+
+from .protein import Protein, describe
+from .database import ProteinDB, ProteinModel, ProteinSimpleModel, read_fasta
+from .query import ProteinQuery
+from .blast import protein_blast, deep_blast
+from .visualize import plot_embeddings
+
+__all__ = [
+    "Protein",
+    "describe",
+    "ProteinDB",
+    "ProteinModel",
+    "ProteinSimpleModel",
+    "read_fasta",
+    "ProteinQuery",
+    "protein_blast",
+    "deep_blast",
+    "plot_embeddings",
+]
